requirements.txt
auth0-python==4.2.0
Authlib==1.2.0
certifi==2023.5.7
colorlog==6.7.0
coverage==7.2.7
cryptography==41.0.1
Django==4.2.1
django-consul==0.4.0
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
factory-boy==3.2.1
Faker==18.11.1
Jinja2==3.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
pytest==7.3.1
pytest-django==4.5.2
python-consul==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rest-framework-auth0==0.6.4
rsa==4.9
sentry-sdk==1.25.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==2.0.2
gunicorn==20.1.0

